+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/engine/TwoStepsBatchRequestCallbackWrapper.kt
@@ -2,6 +2,7 @@ package com.mapbox.search.engine
 
 import com.mapbox.search.ResponseInfo
 import com.mapbox.search.SearchMultipleSelectionCallback
+import com.mapbox.search.SearchRequestException
 import com.mapbox.search.SearchRequestTaskImpl
 import com.mapbox.search.common.assertDebug
 import com.mapbox.search.common.reportRelease
@@ -39,12 +40,11 @@ internal class TwoStepsBatchRequestCallbackWrapper(
 
             try {
                 if (!response.isSuccessful) {
-                    var error = httpErrorsCache.getAndRemove(response.requestID)
-                    if (error != null) {
-                        reportRelease(error)
-                    } else {
-                        error = Exception("Unknown error. Response: $response")
-                    }
+                    val error = httpErrorsCache.getAndRemove(response.requestID)
+                        ?: SearchRequestException(message = response.message, code = response.httpCode)
+
+                    reportRelease(error)
+
                     searchRequestTask.markExecutedAndRunOnCallback(callbackExecutor) {
                         onError(error)
                     }